create pull request template within .github folder
This PR will include the PR template, as well as issue templates for future PRs and issues to add cleanliness and organization to our repository. 

Note: Issue templates will be added shortly as I needed to first create the .github folder directly on github to ensure it recognizes its default folder.